requirements.txt
sanic==23.3.0
sanic-ext==23.3.0
mayim==1.1.0
psycopg-binary==3.1.6
mayim[postgres]
sanic-openapi
psycopg~=3.1.9
json-logging~=1.3.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-grpc